pyproject.toml
 [project]
 name = "group-chat-telegram-ai"
 version = "0.1.0"
 description = "Telegram AI assistant bot (kid-safe) + integration tests"
 readme = "README.md"
 requires-python = ">=3.11"
dependencies = [
 "python-telegram-bot>=21.0",
  "python-dotenv>=1.0.1",
 "httpx>=0.27.0",
]
 
 [project.optional-dependencies]
 dev = [
   "pytest>=8.3.0",
   "pytest-asyncio>=0.24.0",
   "ruff>=0.9.0",
 ]
 
 [build-system]
 requires = ["hatchling>=1.25.0"]
 build-backend = "hatchling.build"
 
[tool.hatch.build.targets.wheel]
packages = ["src/group_chat_telegram_ai"]

 [tool.pytest.ini_options]
 addopts = "-q"
 asyncio_mode = "auto"
 testpaths = ["tests"]
 
 [tool.ruff]
 line-length = 100
 target-version = "py311"
